Updated e2e tests
